requirements.txt
termcolor==1.1.0
GitPython==2.1.1
PyYAML==5.1
